scheduler/sample_drilling_scheduler_app/requirements.txt
corva-sdk==1.0.1
pytest==6.2.4
pydantic==1.8.2
